Add support for resources
CAP requires a MIME type, but it is unlikely the end-user will know what
this is. So, this introduces an internal /api/mime endpoint to fetch a
given URL and return its MIME type (we cannot do this client-side due to
CORS). This has a useful benefit of ensuring the resource is indeed live
and returning a 200.