Add API unit tests
~70% branch coverage so far.

This uses a live PostgreSQL database in Docker to effectively test the
database interactions in the API; mocking would be too brittle and would
not test the API sufficiently.

These tests also helped fix some bugs in the API logic.

Caching of the WhatNow and WMO Register of Alerting Authorities has also
been added.