Implement 'other' Alerting Authority option
This will send an email to the IFRC to verify the AA and set its name.

It can be thought of as becoming an 'IFRC-managed AA'.

In all cases, an AA listed in the WMO Register of AAs should be
preferred, to reduce the administrative load on the IFRC for managing
'custom' AAs.