Re-add WhatNow message integration to single-page Editor UI
Adds a modal dialog showing the different WhatNow messages (event type)
for the current AA country code, the different languages that it has
translations for, and an option to append messages from the various
stages to the Instruction field.

Note: we are only using WhatNow for the Instruction field, not
description or headline or event.