Update Alert Identifier format
We now use [AA ID].[Sent Date].

The DB Alert still uses a UUID as the identifier, because we create
alerts in draft status with a 'sent' time that is different to the final
'sent' time when the alert is published.

So, instead of having to change the DB Alert's ID when an alert is
published (i.e., the 'sent' time is finalised), we use a UUID that stays
the same forever, and set the JSON data identifier to be based on the AA
ID and Sent Time.